requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
jinja2>=3.1.6
python-dotenv>=1.1.1
apscheduler>=3.11.0
pyyaml>=6.0.2
akshare>=1.17.0
